Model/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
sqlalchemy>=2.0.25
pydantic[email]>=2.6.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
facenet-pytorch>=2.5.3
torch>=2.0.0
torchvision>=0.15.0
pillow>=10.0.0
qdrant-client>=1.7.0
python-dotenv>=1.0.0
requests>=2.31.0

